rename virtualize function
